refactor: var field components
