+++ src/components/Services/VariablesDisplay/Fields/ViewField.tsx
