+++ src/components/Services/VariablesDisplay/Fields/index.tsx
@@ -1,1 +1,2 @@
 export * from './ViewField'
+export * from './RawField'
