--- src/components/Services/VariablesDisplay/VariablesList/index.tsx
@@ -1,1 +1,0 @@
-export * from './VariablesList'
